[#21] Feat: 입력 데이터 추가, 모듈 분리
- submit 버튼을 누르면 입력 정보 Map에 정보를 넣게 추가
- 인풋 필드 객체를 밖으로 모듈화
